training_field/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
python-dotenv>=1.0.0
openai>=1.40.0
aiosqlite>=0.19.0
rich>=13.0.0
click>=8.1.0
